+++ clients/admin-ui/cypress/support/commands.ts
@@ -2,8 +2,12 @@
 
 import { STORAGE_ROOT_KEY, USER_PRIVILEGES } from "~/constants";
 
-Cypress.Commands.add("getByTestId", (selector, ...args) =>
-  cy.get(`[data-testid='${selector}']`, ...args)
+Cypress.Commands.add("getByTestId", (selector, options) =>
+  cy.get(`[data-testid='${selector}']`, options)
+);
+
+Cypress.Commands.add("getByTestIdPrefix", (prefix, options) =>
+  cy.get(`[data-testid^='${prefix}']`, options)
 );
 
 Cypress.Commands.add("login", () => {
@@ -34,20 +38,34 @@ Cypress.Commands.add("login", () => {
 
 declare global {
   namespace Cypress {
+    type GetBy = (
+      selector: string,
+      options?: Partial<
+        Cypress.Loggable &
+          Cypress.Timeoutable &
+          Cypress.Withinable &
+          Cypress.Shadow
+      >
+    ) => Chainable<JQuery<HTMLElement>>;
+
     interface Chainable {
       /**
-       * Custom command to select DOM element by data-testid attribute
+       * Custom command to select DOM element by data-testid attribute.
        * @example cy.getByTestId('clear-btn')
        */
-      getByTestId(
-        selector: string,
-        options?: Partial<
-          Cypress.Loggable &
-            Cypress.Timeoutable &
-            Cypress.Withinable &
-            Cypress.Shadow
-        >
-      ): Chainable<JQuery<HTMLElement>>;
+      getByTestId: GetBy;
+      /**
+       * Custom command to select DOM element by the prefix of a data-testid attribute. Useful for
+       * elements that get rendered in a list where each item has its own unique id.
+       *
+       * @example
+       * cy.getByTestIdPrefix('row')
+       * // => [ tr#01, tr#02, ..., tr#20]
+       * // Versus:
+       * cy.getByTestId('row-13')
+       * // => tr#13
+       */
+      getByTestIdPrefix: GetBy;
       /**
        * Programmatically login with a mock user
        */