+++ clients/admin-ui/src/features/privacy-requests/RequestDetails.tsx
@@ -1,12 +1,23 @@
-import { Box, Divider, Flex, Heading, HStack, Tag, Text } from "@fidesui/react";
-import DaysLeftTag from "common/DaysLeftTag";
-import { PrivacyRequestEntity } from "privacy-requests/types";
+import {
+  Box,
+  ButtonGroup,
+  Divider,
+  Flex,
+  Heading,
+  HStack,
+  Tag,
+  Text,
+} from "@fidesui/react";
 
+import ClipboardButton from "~/features/common/ClipboardButton";
+import DaysLeftTag from "~/features/common/DaysLeftTag";
+import RequestStatusBadge from "~/features/common/RequestStatusBadge";
+import RequestType from "~/features/common/RequestType";
+import { PrivacyRequestEntity } from "~/features/privacy-requests/types";
 import { PrivacyRequestStatus as ApiPrivacyRequestStatus } from "~/types/api/models/PrivacyRequestStatus";
 
-import ClipboardButton from "../common/ClipboardButton";
-import RequestStatusBadge from "../common/RequestStatusBadge";
-import RequestType from "../common/RequestType";
+import ApproveButton from "./buttons/ApproveButton";
+import DenyButton from "./buttons/DenyButton";
 import ReprocessButton from "./buttons/ReprocessButton";
 
 type RequestDetailsProps = {
@@ -63,12 +74,23 @@ const RequestDetails = ({ subjectRequest }: RequestDetailsProps) => {
         </Text>
         <HStack spacing="16px">
           <RequestStatusBadge status={status} />
-          {status === "error" && (
-            <ReprocessButton
-              buttonProps={{ size: "xs" }}
-              subjectRequest={subjectRequest}
-            />
-          )}
+          <ButtonGroup isAttached variant="outline" borderRadius="md">
+            {status === "error" && (
+              <ReprocessButton
+                buttonProps={{ size: "xs" }}
+                subjectRequest={subjectRequest}
+              />
+            )}
+
+            {status === "pending" && (
+              <>
+                <ApproveButton subjectRequest={subjectRequest}>
+                  Approve
+                </ApproveButton>
+                <DenyButton subjectRequest={subjectRequest}>Deny</DenyButton>
+              </>
+            )}
+          </ButtonGroup>
 
           <DaysLeftTag
             daysLeft={subjectRequest.days_left}